remove RecipeBookComponent.render()
