+++ common/src/main/java/marsh/town/brb/recipe/BRBRecipeBookCategoryManager.java
@@ -1,0 +1,6 @@
+package marsh.town.brb.recipe;
+
+public class BRBRecipeBookCategoryManager {
+//    public List<? extends GenericRecipeBookCollection> getCollectionsForCategory(BRBCategory category) {
+//    }
+}
